move bigger views to a separate directory
