+++ app/BigImmich/Utils/ThumbnailView.swift
